Made testing script perform in a loop

+++ script.sh
@@ -1,14 +1,38 @@
 #!/bin/bash
-
+MINARGS=1
+if [ $# -lt $MINARGS ]; then
+	echo "Usage: sh script.sh 'flow-one' | 'flow-ten' | 'flow-thousand' | 'big' | 'big-superposition'"
+	exit $E_NOARGS
+fi
+REQ_RESULTS=0
+OWN_RESULTS=0
 ARG=$1
-./generator --$ARG > argFILE
-./lem-in argFILE > outFILE
-OUTNUM=$(wc -l "outFILE" | awk '{print $1}')
-ARGNUM=$(wc -l "argFILE" | awk '{print $1}')
-REQ=$((grep 'require' argFILE | awk '{print $8}') | awk '{print $1}')
-echo $REQ
-NUM=$(($OUTNUM - $ARGNUM - 1))
-#cat "outFILE"
-echo $NUM
-rm -rf outFILE
-rm -rf argFILE
+i=1
+while [ $i -le 20 ]
+do
+	./generator --$ARG > argFILE
+	./lem-in argFILE > outFILE
+	OUTNUM=$(wc -l "outFILE" | awk '{print $1}')
+	ARGNUM=$(wc -l "argFILE" | awk '{print $1}')
+	REQ=$(head -5 'argFILE' |  grep 'require' | awk '{print $8}')
+	REQ_RESULTS=$(( $REQ_RESULTS + $REQ ))
+	LINES_NUM=$(( $OUTNUM - $ARGNUM - 1 ))
+	OWN_RESULTS=$(( $OWN_RESULTS + $LINES_NUM ))
+	echo "req $REQ"
+	echo "own $LINES_NUM"
+	if (( $LINES_NUM < 0 ))
+	then
+		cp -rf outFILE outProblem
+		cp -rf argFILE inProblem
+	fi
+	rm -rf outFILE
+	rm -rf argFILE
+	i=$(( $i + 1 ))
+done
+i=$(( $i - 1))
+echo $i
+REQ_AVERAGE=$(( $REQ_RESULTS / $i))
+OWN_AVERAGE=$(( $OWN_RESULTS / $i))
+echo "Required average per $i tests: $REQ_AVERAGE"
+echo "Our average per $i tests: $OWN_AVERAGE"
+exit 0
